requirements.txt
# Reinforcement Learning Environment Dependencies
gymnasium[box2d]==0.29.1
torch>=2.6.0
numpy>=1.24.0
tqdm>=4.65.0
tensorboard>=2.15.0

# Optional: For development and testing
pytest>=7.4.0
black>=23.0.0

# yaml
pyyaml>=6.0